requirements.txt
django 
djangorestframework 
psycopg2-binary 
pika 
python-dotenv 
celery 
redis 
requests
django-celery-results
django-celery-beat